FIX: ServerCommunicator

+++ src/main/java/de/goldendeveloper/entertainment/discord/Events.java
@@ -224,7 +224,9 @@ public class Events extends ListenerAdapter {
             if (e.getMember().hasPermission(Permission.ADMINISTRATOR)) {
                 if (CountingGame.hasCountingGame(e.getGuild())) {
                     CountingGame countingGame = CountingGame.findGame(Objects.requireNonNull(e.getGuild()));
-                    countingGame.delete();
+                    if (countingGame != null) {
+                        countingGame.delete();
+                    }
                     e.reply("Der Counting Channel wurde gelöscht!").queue();
                 } else {
                     e.reply("Dieser Server keinen  Zähl Channel bitte erstelle zuerst einen mit dem Befehl: /" + Discord.cmdCountingGame + "!").queue();
@@ -301,13 +303,13 @@ public class Events extends ListenerAdapter {
 
     @Override
     public void onGuildJoin(GuildJoinEvent e) {
-        Main.getServerCommunicator().sendToServer(ServerCommunicator.action.ADD, e.getGuild().getId());
+        Main.getServerCommunicator().addServer(e.getGuild().getId());
         e.getJDA().getPresence().setActivity(Activity.playing("/help | " + e.getJDA().getGuilds().size() + " Servern"));
     }
 
     @Override
     public void onGuildLeave(GuildLeaveEvent e) {
-        Main.getServerCommunicator().sendToServer(ServerCommunicator.action.REMOVE, e.getGuild().getId());
+        Main.getServerCommunicator().removeServer(e.getGuild().getId());
         e.getJDA().getPresence().setActivity(Activity.playing("/help | " + e.getJDA().getGuilds().size() + " Servern"));
     }
